moved elemental NPC stats to a new ElementalNPC class
